Refactor auto login into custom hook
